ADD：用户积分兑换订单下单；管理员更新发货时间；用户确认收货接口

+++ src/main/java/com/trash/collection/trash/product/domain/Product.java
@@ -70,7 +70,7 @@ public class Product implements Serializable {
      * 兑换所需积分
      */
     @TableField("need_points")
-    private Integer needPoints;
+    private BigDecimal needPoints;
     /**
      * 商品总数
      */
